CHANGE Use case 4 bug fixes

+++ nginx-in-docker/docker-stop-rm.sh
@@ -1,6 +1,8 @@
 #/bin/sh
 set -e
 containers_to_rm_list=$(docker ps -a -q --filter ancestor="$1" --format="{{.ID}}")
-echo $containers_to_rm_list | xargs -L1 docker stop
-echo $containers_to_rm_list | xargs -L1 docker rm 
+if [[ -n "$containers_to_rm_list" ]]; then
+  echo $containers_to_rm_list | xargs -L1 docker stop
+  echo $containers_to_rm_list | xargs -L1 docker rm 
+fi
